% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Annotation.r
\name{Annotation}
\alias{Annotation}
\title{Class representing annotation for expression data.}
\description{
An R6 class to represent transcriptome annotation with protein mapping.
Specifically, it provides the following features:
\itemize{
\item Transcriptome annotation:
\itemize{
\item Association of transcript IDs with gene IDs, RNA types, and species
\item Association of gene IDs with gene symbols, protein IDs (UniProt), etc.
}
}
}
\details{
For more information on how to specify the arguments, please refer to the
vignette defining file formats.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Annotation-new}{\code{Annotation$new()}}
\item \href{#method-Annotation-generate_translate_dict}{\code{Annotation$generate_translate_dict()}}
\item \href{#method-Annotation-get_from_ids}{\code{Annotation$get_from_ids()}}
\item \href{#method-Annotation-get_to_ids}{\code{Annotation$get_to_ids()}}
\item \href{#method-Annotation-clone}{\code{Annotation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Annotation-new"></a>}}
\if{latex}{\out{\hypertarget{method-Annotation-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new \code{Annotation} object.

This method initializes a new \code{Annotation} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Annotation$new(annotation = NULL, format_gff = TRUE, idmapping = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{annotation}}{Path to the annotation file in GFF format / or as described in the vignette.
can be a vector of files.}

\item{\code{format_gff}}{If \code{TRUE}, the annotation file is in GFF format.}

\item{\code{idmapping}}{Path to the idmapping file, which allows the
mapping of uniprot id to gene id. If not give, the idmapping file will be
downloaded from uniprot using taxon ID from annotation file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Annotation} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Annotation-generate_translate_dict"></a>}}
\if{latex}{\out{\hypertarget{method-Annotation-generate_translate_dict}{}}}
\subsection{Method \code{generate_translate_dict()}}{
generate a translate dictionary (named vector) from and id to another
based on the annotation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Annotation$generate_translate_dict(from, to)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{from}}{the id type corresponding to the keys}

\item{\code{to}}{the id type corresponding to the values}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Below is precised what is available and id type for value depending
on id type for keys
\itemize{
\item Entrez Gene ID \code{gid} (Or equivalent on Ensembl,UCSC...)
\itemize{
\item Gene symbol \code{symbol}
\item Protein name \code{protein_names}
\item Protein ID \code{uniprot}
\item Taxon ID \code{tax_id}
\item Taxon name \code{tax_name}
\item Typed gene ID \code{tgid}
}
\item RefSeq transcript ID \code{txid} (Or equivalent on Ensembl,UCSC...)
\itemize{
\item Entrez Gene ID \code{gid}
\item rna type \code{type}
\item Taxon ID \code{tax_id}
\item Taxon name \code{tax_name}
\item Typed gene ID \code{tgid}
}
\item Typed gene ID \code{tgid} i.e. gene id with type concatenated (sep is _)
(one gene can have several type, some coding, some not coding)
\itemize{
\item Gene symbol \code{symbol}
\item Protein name \code{protein_names} (warning even if not coding type)
\item Protein ID \code{uniprot} (warning even if not coding type)
\item rna type \code{type}
\item Taxon ID \code{tax_id}
\item Taxon name \code{tax_name}
}
\item Uniprot protein ID \code{uniprot}
\itemize{
\item Entrez Gene ID \code{gid}
\item Protein name \code{protein_names}
\item Gene symbol \code{symbol}
}
}
}

\subsection{Returns}{
a named vector that can translate from an id type to another
Get the list of id type available for translate dicttionary as keys
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Annotation-get_from_ids"></a>}}
\if{latex}{\out{\hypertarget{method-Annotation-get_from_ids}{}}}
\subsection{Method \code{get_from_ids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Annotation$get_from_ids()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
vecotr with potentioal key "from"
Get the list of id type available for translate dicttionary as values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Annotation-get_to_ids"></a>}}
\if{latex}{\out{\hypertarget{method-Annotation-get_to_ids}{}}}
\subsection{Method \code{get_to_ids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Annotation$get_to_ids(from)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{from}}{the key of the translate dictionary}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vecotr with potentioal key "to"
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Annotation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Annotation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Annotation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
