% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_plot_complex.r
\name{arrange_complex_plots}
\alias{arrange_complex_plots}
\title{Helper function to arrange complex plots}
\usage{
arrange_complex_plots(graphs)
}
\arguments{
\item{graphs}{List of ggplot objects}
}
\value{
Arranged plot grid
}
\description{
Helper function to arrange complex plots
}
\keyword{internal}
