% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_norm.r
\name{calc_norm_fac}
\alias{calc_norm_fac}
\title{Helper function for median normalization factor calculation}
\usage{
calc_norm_fac(tgt_vector, ref_vector, norm_mean_fun, trim_extreme)
}
\arguments{
\item{tgt_vector}{target expression vector}

\item{ref_vector}{reference expression vector}

\item{mean_fun}{mean function}
}
\value{
normalization factors
}
\description{
Helper function for median normalization factor calculation
}
\keyword{internal}
