% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_norm.r
\name{calc_norm_fact_median}
\alias{calc_norm_fact_median}
\title{Helper function for median normalization factor calculation}
\usage{
calc_norm_fact_median(norm, sample_ref)
}
\arguments{
\item{norm}{normalized matrix}

\item{sample_ref}{reference sample values}
}
\value{
normalization factors
}
\description{
Helper function for median normalization factor calculation
}
