% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_norm.r
\name{compute_inter_norm}
\alias{compute_inter_norm}
\title{Helper function to do inter normalization}
\usage{
compute_inter_norm(
  raw,
  norm_fact,
  ref_type,
  ref_samples,
  design,
  norm_scale,
  norm_by,
  ref_mean,
  norm_mean,
  tgt_mean,
  a_mean,
  a_trim_value,
  m_trim_prop,
  trim_extreme = FALSE
)
}
\arguments{
\item{raw}{raw matrix}

\item{norm_fact}{intra norm fact matrix}

\item{ref_type}{reference type}

\item{ref_samples}{reference samples}

\item{design}{private$design}

\item{norm_scale}{scale of normalization}

\item{norm_by}{normalization by group or sample}

\item{ref_mean}{mean function for reference}

\item{norm_mean}{mean function for normalization}

\item{tgt_mean}{mean function for target}

\item{a_mean}{mean function for A}

\item{a_trim_value}{trim value for A}

\item{m_trim_prop}{trim proportion for M}
}
\value{
inter norm fact matrix
}
\description{
Helper function to do inter normalization
}
\keyword{internal}
