% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_norm.r
\name{compute_intra_norm_factor}
\alias{compute_intra_norm_factor}
\title{Helper function for ExprData$compute_and_set_intra_norm_fact}
\usage{
compute_intra_norm_factor(method, selected_ids, raw_matrix, len_matrix)
}
\arguments{
\item{method}{see ExprData$compute_and_set_intra_norm_fact}

\item{selected_ids}{private$selected_ids (see ExprData)}

\item{raw_matrix}{raw count matrix}

\item{len_matrix}{length matrix}
}
\value{
intra normalization factor matrix
}
\description{
Helper function for ExprData$compute_and_set_intra_norm_fact
}
