% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_plot_complex.r
\name{create_corr_plot}
\alias{create_corr_plot}
\title{Helper function to create correlation plots}
\usage{
create_corr_plot(in_data, in_title, ggplot_mod = NULL)
}
\arguments{
\item{in_data}{Expression data matrix}

\item{in_title}{Plot title}

\item{ggplot_mod}{Additional ggplot modifications}
}
\value{
ggplot object (ggmatrix)
}
\description{
Helper function to create correlation plots
}
