% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_plot_complex.r
\name{create_hclust_plot}
\alias{create_hclust_plot}
\title{Helper function to create hierarchical clustering plots}
\usage{
create_hclust_plot(
  in_data,
  in_title,
  data_design,
  dist_method = "euclidean",
  hclust_method = "ward.D2",
  dim_reduce = NULL,
  clust_bar_var = c(),
  height_main = 10,
  width_main = 4,
  color_palette = "BrBg",
  prcomp_args = list(),
  ggplot_mod = NULL
)
}
\arguments{
\item{in_data}{Expression data matrix}

\item{in_title}{Plot title}

\item{data_design}{Design data frame}

\item{dist_method}{Distance method}

\item{hclust_method}{Hierarchical clustering method}

\item{dim_reduce}{Dimensionality reduction parameter}

\item{clust_bar_var}{Variables for color bars}

\item{height_main}{Main plot height}

\item{width_main}{Main plot width}

\item{color_palette}{Color palette name}

\item{prcomp_args}{Arguments for prcomp}

\item{ggplot_mod}{Additional ggplot modifications}
}
\value{
ggplot object or arranged plots
}
\description{
Helper function to create hierarchical clustering plots
}
