% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_plot_complex.r
\name{create_prcomp_plot}
\alias{create_prcomp_plot}
\title{Helper function to create PCA plots}
\usage{
create_prcomp_plot(
  in_data,
  in_title,
  data_design,
  prcomp_args = list(),
  prcomp_autoplot_args = list(),
  ggplot_mod = NULL,
  color_palette = "BrBg"
)
}
\arguments{
\item{in_data}{Expression data matrix}

\item{in_title}{Plot title}

\item{data_design}{Design data frame}

\item{prcomp_args}{Arguments for prcomp}

\item{prcomp_autoplot_args}{Arguments for autoplot}

\item{ggplot_mod}{Additional ggplot modifications}

\item{color_palette}{Color palette name}
}
\value{
ggplot object
}
\description{
Helper function to create PCA plots
}
