% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Annotation.r
\name{create_uniprot_gene_id_mappings}
\alias{create_uniprot_gene_id_mappings}
\title{utils function for this class only
scores min given for unique first and then for dups before merging
switch to data.table}
\usage{
create_uniprot_gene_id_mappings(
  idmapping,
  unique_reviewed_min_score = 1L,
  dups_reviewed_min_score = 4L,
  unique_unreviewed_min_score = 2L,
  dups_unreviewed_min_score = 5L
)
}
\arguments{
\item{idmapping}{data.table with id mappings from uniprot KB}

\item{annotation}{data.table with annotation from gff files}

\item{reviewed_min_scores}{vector of two integers with min scores for
reviewed entries, first for unique and then for dups}

\item{unreviewed_min_scores}{vector of two integers with min scores for
unreviewed entries, first for unique and then for dups}
}
\description{
utils function for this class only
scores min given for unique first and then for dups before merging
switch to data.table
}
