% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\docType{data}
\name{DOWN_COLOR}
\alias{DOWN_COLOR}
\title{Color palette for nexomis: Down blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
DOWN_COLOR
}
\description{
Color palette for nexomis: Down blue
}
\keyword{datasets}
