% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData.r
\name{ExprDataTranscript}
\alias{ExprDataTranscript}
\title{Class representing an expression dataset at transcript-level}
\description{
Extension of expression dataset at transcript-level
}
\section{Super class}{
\code{\link[nexodiff:ExprData]{nexodiff::ExprData}} -> \code{ExprDataTranscript}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExprDataTranscript-new}{\code{ExprDataTranscript$new()}}
\item \href{#method-ExprDataTranscript-clone}{\code{ExprDataTranscript$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="clean_id_versions"><a href='../../nexodiff/html/ExprData.html#method-ExprData-clean_id_versions'><code>nexodiff::ExprData$clean_id_versions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="compute_and_set_inter_norm_fact"><a href='../../nexodiff/html/ExprData.html#method-ExprData-compute_and_set_inter_norm_fact'><code>nexodiff::ExprData$compute_and_set_inter_norm_fact()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="compute_and_set_intra_norm_fact"><a href='../../nexodiff/html/ExprData.html#method-ExprData-compute_and_set_intra_norm_fact'><code>nexodiff::ExprData$compute_and_set_intra_norm_fact()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="compute_norm"><a href='../../nexodiff/html/ExprData.html#method-ExprData-compute_norm'><code>nexodiff::ExprData$compute_norm()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="compute_norm_fact"><a href='../../nexodiff/html/ExprData.html#method-ExprData-compute_norm_fact'><code>nexodiff::ExprData$compute_norm_fact()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="extract_pairwise_data_with_design"><a href='../../nexodiff/html/ExprData.html#method-ExprData-extract_pairwise_data_with_design'><code>nexodiff::ExprData$extract_pairwise_data_with_design()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="filter_and_get_len"><a href='../../nexodiff/html/ExprData.html#method-ExprData-filter_and_get_len'><code>nexodiff::ExprData$filter_and_get_len()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="filter_and_get_raw"><a href='../../nexodiff/html/ExprData.html#method-ExprData-filter_and_get_raw'><code>nexodiff::ExprData$filter_and_get_raw()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="filter_and_set_selected_ids"><a href='../../nexodiff/html/ExprData.html#method-ExprData-filter_and_set_selected_ids'><code>nexodiff::ExprData$filter_and_set_selected_ids()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="get_annotation"><a href='../../nexodiff/html/ExprData.html#method-ExprData-get_annotation'><code>nexodiff::ExprData$get_annotation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="get_design"><a href='../../nexodiff/html/ExprData.html#method-ExprData-get_design'><code>nexodiff::ExprData$get_design()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="get_inter_norm_fact_opts"><a href='../../nexodiff/html/ExprData.html#method-ExprData-get_inter_norm_fact_opts'><code>nexodiff::ExprData$get_inter_norm_fact_opts()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="get_intra_norm_fact_method"><a href='../../nexodiff/html/ExprData.html#method-ExprData-get_intra_norm_fact_method'><code>nexodiff::ExprData$get_intra_norm_fact_method()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="get_len"><a href='../../nexodiff/html/ExprData.html#method-ExprData-get_len'><code>nexodiff::ExprData$get_len()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="get_main_etag"><a href='../../nexodiff/html/ExprData.html#method-ExprData-get_main_etag'><code>nexodiff::ExprData$get_main_etag()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="get_raw"><a href='../../nexodiff/html/ExprData.html#method-ExprData-get_raw'><code>nexodiff::ExprData$get_raw()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="is_at_gene_level"><a href='../../nexodiff/html/ExprData.html#method-ExprData-is_at_gene_level'><code>nexodiff::ExprData$is_at_gene_level()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="is_with_fixed_length"><a href='../../nexodiff/html/ExprData.html#method-ExprData-is_with_fixed_length'><code>nexodiff::ExprData$is_with_fixed_length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="plot_corr"><a href='../../nexodiff/html/ExprData.html#method-ExprData-plot_corr'><code>nexodiff::ExprData$plot_corr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="plot_dist_per_sample"><a href='../../nexodiff/html/ExprData.html#method-ExprData-plot_dist_per_sample'><code>nexodiff::ExprData$plot_dist_per_sample()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="plot_hclust"><a href='../../nexodiff/html/ExprData.html#method-ExprData-plot_hclust'><code>nexodiff::ExprData$plot_hclust()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="plot_prcomp"><a href='../../nexodiff/html/ExprData.html#method-ExprData-plot_prcomp'><code>nexodiff::ExprData$plot_prcomp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="plot_sum_per_type_per_sample"><a href='../../nexodiff/html/ExprData.html#method-ExprData-plot_sum_per_type_per_sample'><code>nexodiff::ExprData$plot_sum_per_type_per_sample()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="reset"><a href='../../nexodiff/html/ExprData.html#method-ExprData-reset'><code>nexodiff::ExprData$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="show_etags_summary"><a href='../../nexodiff/html/ExprData.html#method-ExprData-show_etags_summary'><code>nexodiff::ExprData$show_etags_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nexodiff" data-topic="ExprData" data-id="sum_per_type_per_sample"><a href='../../nexodiff/html/ExprData.html#method-ExprData-sum_per_type_per_sample'><code>nexodiff::ExprData$sum_per_type_per_sample()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExprDataTranscript-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExprDataTranscript-new}{}}}
\subsection{Method \code{new()}}{
Initialize \code{ExprDataTranscript} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExprDataTranscript$new(
  design,
  annotation,
  with_fixed_length = FALSE,
  log_level = "WARN",
  format = "kallisto",
  suffix_pattern = "\\\\.\\\\d+$"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{design}}{A pairwise design object \link{PairwiseDesign}}

\item{\code{annotation}}{An annotation object \link{Annotation}}

\item{\code{with_fixed_length}}{A boolean indicating whether the NGS library is
of a fixed length like 3 prime or not}

\item{\code{log_level}}{Level of logging (see logging package). Default = WARN}

\item{\code{format}}{Format of files. Default=kallisto}

\item{\code{suffix_pattern}}{See clean_txid_versions method
Default = txid}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ExprDataTranscript} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExprDataTranscript-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExprDataTranscript-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExprDataTranscript$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
