% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_extract.r
\name{extract_pairwise_data}
\alias{extract_pairwise_data}
\title{Helper function for ExprData$extract_pairwise_data_with_design}
\usage{
extract_pairwise_data(
  design,
  raw_matrix,
  len_matrix,
  intra_norm_fact_matrix,
  inter_norm_fact,
  in_batch,
  in_group,
  include_ctrl = TRUE
)
}
\arguments{
\item{design}{private$design (see ExprData)}

\item{raw_matrix}{raw count matrix}

\item{len_matrix}{length matrix}

\item{intra_norm_fact_matrix}{intra normalization factor matrix}

\item{inter_norm_fact}{inter normalization factors}

\item{in_batch}{see ExprData$extract_pairwise_data_with_design}

\item{in_group}{see ExprData$extract_pairwise_data_with_design}

\item{include_ctrl}{see ExprData$extract_pairwise_data_with_design}
}
\value{
list with selected attributes (see ExprData$extract_pairwise
}
\description{
Helper function for ExprData$extract_pairwise_data_with_design
}
\keyword{data_with_design)}
\keyword{internal}
