% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_filter.r
\name{filter_etags}
\alias{filter_etags}
\title{Helper function for ExprData$filter_and_set_selected_ids}
\usage{
filter_etags(
  current_selected_ids,
  annotation,
  main_etag,
  values,
  filtered_var = "type",
  filter_type = "keep"
)
}
\arguments{
\item{current_selected_ids}{private$selected_ids (see ExprData)}

\item{annotation}{private$annotation (see ExprData)}

\item{main_etag}{see ExprData$filter_and_set_selected_ids}

\item{values}{see ExprData$filter_and_set_selected_ids}

\item{filtered_var}{see ExprData$filter_and_set_selected_ids}

\item{filter_type}{see ExprData$filter_and_set_selected_ids}
}
\value{
Vector of filtered expression tag IDs
}
\description{
Helper function for ExprData$filter_and_set_selected_ids
}
