% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_norm.r
\name{get_reference}
\alias{get_reference}
\title{Helper function to get reference for normalization}
\usage{
get_reference(norm, ctrl_samples, norm_ref, norm_ref_mean_fun)
}
\arguments{
\item{norm}{normalized matrix}

\item{ctrl_samples}{control sample names}

\item{norm_ref}{reference specification}

\item{norm_ref_mean_fun}{mean function for reference calculation}
}
\value{
reference values
}
\description{
Helper function to get reference for normalization
}
