% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_norm.r
\name{get_reference_ids}
\alias{get_reference_ids}
\title{Helper function to get reference ids}
\usage{
get_reference_ids(
  ids,
  ctrl_samples,
  ref_type = c("all", "ctrl", "specified"),
  ref_samples = NULL
)
}
\arguments{
\item{ids}{possible sample ids}

\item{ctrl_samples}{control sample ids}

\item{ref_type}{type of reference
\itemize{
\item "all": use all samples as reference
\item "ctrl": use control samples as reference
\item "specified": use specified samples as reference
}}

\item{ref_samples}{specified reference samples}
}
\description{
Helper function to get reference ids
}
\keyword{internal}
