% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_norm.r
\name{get_trimmed_ids}
\alias{get_trimmed_ids}
\title{Helper function to get trimmed ids}
\usage{
get_trimmed_ids(trim_ref, trim_grp, mean_fun, a_trim_value, m_trim_prop)
}
\arguments{
\item{trim_ref}{reference expression vector}

\item{trim_grp}{expression vector}

\item{mean_fun}{mean function to compute average}

\item{a_trim_value}{trim value}

\item{m_trim_prop}{trim proportion top and bottom}
}
\value{
trimmed ids
}
\description{
Helper function to get trimmed ids
}
\keyword{internal}
