% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_norm.r
\name{get_trimmed_mat}
\alias{get_trimmed_mat}
\title{Helper function to get trimmed matrix}
\usage{
get_trimmed_mat(
  raw,
  norm_fact,
  a_trim_norm,
  a_trim_value,
  a_trim_mean_fun,
  replace_zero_by
)
}
\arguments{
\item{raw}{raw count matrix}

\item{norm_fact}{normalization factors}

\item{a_trim_norm}{whether to use normalized values for trimming}

\item{a_trim_value}{threshold for trimming}

\item{a_trim_mean_fun}{mean function for trimming}

\item{replace_zero_by}{value to replace zeros}
}
\value{
trimmed matrix
}
\description{
Helper function to get trimmed matrix
}
