% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_norm.r
\name{inter_norm_check}
\alias{inter_norm_check}
\title{Helper function to check args}
\usage{
inter_norm_check(
  method,
  norm_scale,
  norm_by,
  norm_ref,
  norm_ref_mean,
  m_trim_prop,
  m_trim_mean,
  a_trim_value,
  a_trim_norm,
  a_trim_mean,
  replace_zero_by,
  ncpus
)
}
\arguments{
\item{method}{see ExprData$compute_and_set_intra_norm_fact}

\item{norm_scale}{see ExprData$compute_and_set_intra_norm_fact}

\item{norm_by}{see ExprData$compute_and_set_intra_norm_fact}

\item{norm_ref}{see ExprData$compute_and_set_intra_norm_fact}

\item{norm_ref_mean}{see ExprData$compute_and_set_intra_norm_fact}

\item{m_trim_prop}{see ExprData$compute_and_set_intra_norm_fact}

\item{m_trim_mean}{see ExprData$compute_and_set_intra_norm_fact}

\item{a_trim_value}{see ExprData$compute_and_set_intra_norm_fact}

\item{a_trim_norm}{see ExprData$compute_and_set_intra_norm_fact}

\item{a_trim_mean}{see ExprData$compute_and_set_intra_norm_fact}

\item{replace_zero_by}{see ExprData$compute_and_set_intra_norm_fact}

\item{ncpus}{see ExprData$compute_and_set_intra_norm_fact}
}
\description{
Helper function to check args
}
