% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_norm.r
\name{main_calc_norm_fac}
\alias{main_calc_norm_fac}
\title{Main helper function for normalization factor calculation}
\usage{
main_calc_norm_fac(
  mat,
  ref_mat,
  ref_mean_fun,
  norm_mean_fun,
  tgt_mean_fun,
  a_mean_fun,
  a_trim_value,
  m_trim_prop,
  grouped_samples,
  trim_extreme
)
}
\arguments{
\item{mat}{expression matrix}

\item{ref_mat}{reference expression matrix}

\item{ref_mean_fun}{mean function for reference}

\item{norm_mean_fun}{mean function for normalization}

\item{tgt_mean_fun}{mean function for target}

\item{a_mean_fun}{mean function for A vector}

\item{a_trim_value}{trim value for A vector}

\item{m_trim_prop}{trim proportion for M vector}

\item{grouped_samples}{grouped samples}
}
\description{
Main helper function for normalization factor calculation
}
\keyword{internal}
