% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_test_data.r
\name{make_test_data}
\alias{make_test_data}
\title{Simulate data from config}
\usage{
make_test_data(set_id, tmp_dir, quant_source = "kallisto")
}
\arguments{
\item{set_id}{set_id for data simulation}

\item{tmp_dir}{where to create the dataset}

\item{quant_source}{The source of quantification files. Either "kallisto"
(default) or "salmon". This determines the output file format.}
}
\value{
create all files and return paths
}
\description{
Data are generated in a temporary location and paths are returned
}
