% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{make_test_data_from_xlsx}
\alias{make_test_data_from_xlsx}
\title{Simulate data from config}
\usage{
make_test_data_from_xlsx(set_id)
}
\arguments{
\item{set_id}{set_id for data simulation}
}
\value{
create all files and return paths
}
\description{
Data are generated in a temporary location and paths are returned
}
