% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\docType{data}
\name{MID_COLOR}
\alias{MID_COLOR}
\title{Color palette for nexomis: Mid yellow-white}
\format{
An object of class \code{character} of length 1.
}
\usage{
MID_COLOR
}
\description{
Color palette for nexomis: Mid yellow-white
}
\keyword{datasets}
