% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexodiff-package.R
\docType{package}
\name{nexodiff-package}
\alias{nexodiff}
\alias{nexodiff-package}
\title{nexodiff: RNA-seq Differential Expression Analysis}
\description{
Provides classes and functions for performing differential expression analysis on RNA-seq data. The package implements various statistical methods and visualization tools for analyzing gene expression differences between experimental conditions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nexomis/nexodiff}
  \item \url{https://nexomis.github.io/nexodiff/}
  \item Report bugs at \url{https://github.com/nexomis/nexodiff/issues}
}

}
\author{
\strong{Maintainer}: Julien FOURET \email{jfouret@nexomis.io}

Authors:
\itemize{
  \item Abdelghafar Goumaidi
  \item Catalina Gonzalez Gomez
}

}
\keyword{internal}
