% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_PairwiseDesign.r
\name{PairwiseDesign}
\alias{PairwiseDesign}
\title{Class representing metadata for expression data: Design}
\description{
An R6 class to represent metadata for expression data. It combines a pairwise
design of experiments with a 2-level structure and transcriptome annotation.
Specifically, it provides the following features:
\itemize{
\item Pairwise design of experiment with a 2-level structure:
\itemize{
\item Samples are regrouped into \emph{groups}
\item Groups are regrouped into \emph{batches}, where one group is designed as the
control group used as a reference for pairwise comparisons
}
}
}
\details{
For more information on how to specify the arguments, please refer to the
vignette defining file formats.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PairwiseDesign-new}{\code{PairwiseDesign$new()}}
\item \href{#method-PairwiseDesign-get_b_labels}{\code{PairwiseDesign$get_b_labels()}}
\item \href{#method-PairwiseDesign-get_g_labels}{\code{PairwiseDesign$get_g_labels()}}
\item \href{#method-PairwiseDesign-build_file_paths}{\code{PairwiseDesign$build_file_paths()}}
\item \href{#method-PairwiseDesign-extract_sample_names}{\code{PairwiseDesign$extract_sample_names()}}
\item \href{#method-PairwiseDesign-list_groups_per_batches}{\code{PairwiseDesign$list_groups_per_batches()}}
\item \href{#method-PairwiseDesign-list_batches}{\code{PairwiseDesign$list_batches()}}
\item \href{#method-PairwiseDesign-find_control_group_per_batches}{\code{PairwiseDesign$find_control_group_per_batches()}}
\item \href{#method-PairwiseDesign-get_pairwise_design}{\code{PairwiseDesign$get_pairwise_design()}}
\item \href{#method-PairwiseDesign-list_paired_id_per_sample}{\code{PairwiseDesign$list_paired_id_per_sample()}}
\item \href{#method-PairwiseDesign-filter_and_set_selected_samples}{\code{PairwiseDesign$filter_and_set_selected_samples()}}
\item \href{#method-PairwiseDesign-set_selected_samples}{\code{PairwiseDesign$set_selected_samples()}}
\item \href{#method-PairwiseDesign-is_paired}{\code{PairwiseDesign$is_paired()}}
\item \href{#method-PairwiseDesign-reset}{\code{PairwiseDesign$reset()}}
\item \href{#method-PairwiseDesign-clone}{\code{PairwiseDesign$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesign-new"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesign-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new \code{PairwiseDesign} object.

This method initializes a new \code{PairwiseDesign} object.
The object represents metadata for expression data, including pairwise
design of experiments and transcriptome annotation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesign$new(
  pairwise_design_file,
  annotation_file,
  idmapping_file,
  src_dir = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pairwise_design_file}}{Path to the pairwise design file, which must
be in .yml or .csv format.}

\item{\code{src_dir}}{Path(s) specified only in case of csv file with the path
for each run to the folder with expression files. In case of several runs
defined in the design file, a named vector is required. Can be specified
for YML also but only 1 value for all runs.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{PairwiseDesignWithAnnotation} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesign-get_b_labels"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesign-get_b_labels}{}}}
\subsection{Method \code{get_b_labels()}}{
get batch labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesign$get_b_labels()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
named vector with batch label (batch code as names)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesign-get_g_labels"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesign-get_g_labels}{}}}
\subsection{Method \code{get_g_labels()}}{
get group labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesign$get_g_labels()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
named vector with group label (group code as names)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesign-build_file_paths"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesign-build_file_paths}{}}}
\subsection{Method \code{build_file_paths()}}{
get the paths of files with raw expression data for all samples
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesign$build_file_paths()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
named vector with paths and sample name as keys
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesign-extract_sample_names"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesign-extract_sample_names}{}}}
\subsection{Method \code{extract_sample_names()}}{
get sample names with simple design-based filtering
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesign$extract_sample_names(
  in_batch = NULL,
  in_group = NULL,
  in_run = NULL,
  basename = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{in_batch}}{A character vector of batch codes. Select or report only
samples in those batches.}

\item{\code{in_group}}{A character vector of group codes. Select or report only
samples in those groups.}

\item{\code{in_run}}{A character vector of run IDs. Select or report only samples
in those runs.}

\item{\code{basename}}{whether to return the basename in place of the complete
name (with run id)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sample names in character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesign-list_groups_per_batches"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesign-list_groups_per_batches}{}}}
\subsection{Method \code{list_groups_per_batches()}}{
Get a the groups foreach batch
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesign$list_groups_per_batches(include_ctrl = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_ctrl}}{whether to include control group}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list of vectors with batch as key and groups as value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesign-list_batches"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesign-list_batches}{}}}
\subsection{Method \code{list_batches()}}{
Get a all batchs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesign$list_batches()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
vector with batch code
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesign-find_control_group_per_batches"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesign-find_control_group_per_batches}{}}}
\subsection{Method \code{find_control_group_per_batches()}}{
get the control group per batch
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesign$find_control_group_per_batches()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
get named vector with batch as key and ctrl group as value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesign-get_pairwise_design"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesign-get_pairwise_design}{}}}
\subsection{Method \code{get_pairwise_design()}}{
get the pairwise_design table with design-based filtering
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesign$get_pairwise_design(
  in_batch = NULL,
  in_group = NULL,
  in_run = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{in_batch}}{A character vector of batch codes. Select or report only
samples in those batches.}

\item{\code{in_group}}{A character vector of group codes. Select or report only
samples in those groups.}

\item{\code{in_run}}{A character vector of run IDs. Select or report only samples
in those runs.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
pairwise_design data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesign-list_paired_id_per_sample"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesign-list_paired_id_per_sample}{}}}
\subsection{Method \code{list_paired_id_per_sample()}}{
get the id to be used for paired analysis per sample
only return for samples in batch with a paired design
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesign$list_paired_id_per_sample()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
named character vector sample_name -> paired_id
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesign-filter_and_set_selected_samples"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesign-filter_and_set_selected_samples}{}}}
\subsection{Method \code{filter_and_set_selected_samples()}}{
Select samples based on design. The results will be an intersect with
the previous selection. You can reset the object if it's not desired.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesign$filter_and_set_selected_samples(
  in_batch = NULL,
  in_group = NULL,
  in_run = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{in_batch}}{A character vector of batch codes. Select or report only
samples in those batches.}

\item{\code{in_group}}{A character vector of group codes. Select or report only
samples in those groups.}

\item{\code{in_run}}{A character vector of run IDs. Select or report only samples
in those runs.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector of the selected samples
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesign-set_selected_samples"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesign-set_selected_samples}{}}}
\subsection{Method \code{set_selected_samples()}}{
Set selected samples. The results will be an intersect with
the previous selection. You can reset the object if it's not desired.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesign$set_selected_samples(sample_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample_list}}{vector of sample codes to select}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector of the selected samples
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesign-is_paired"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesign-is_paired}{}}}
\subsection{Method \code{is_paired()}}{
get the list of batch with a paired design (DEPRECATED)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesign$is_paired()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesign-reset"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesign-reset}{}}}
\subsection{Method \code{reset()}}{
Reset object, this affects:
\itemize{
\item samples selection (as initially)
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesign$reset()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
vector of the selected samples
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesign-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesign-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesign$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
