% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_PairwiseDesignWithAnnotation.r
\name{PairwiseDesignWithAnnotation}
\alias{PairwiseDesignWithAnnotation}
\title{Class representing metadata for expression data: Design and Gene annotation}
\description{
An R6 class to represent metadata for expression data. It combines a pairwise
design of experiments with a 2-level structure and transcriptome annotation.
Specifically, it provides the following features:
\itemize{
\item Pairwise design of experiment with a 2-level structure:
\itemize{
\item Samples are regrouped into \emph{groups}
\item Groups are regrouped into \emph{batches}, where one group is designed as the
control group used as a reference for pairwise comparisons
}
\item Transcriptome annotation:
\itemize{
\item Association of transcript IDs with gene IDs, RNA types, and species
\item Association of gene IDs with gene symbols, protein IDs (UniProt), etc.
}
}
}
\details{
For more information on how to specify the arguments, please refer to the
vignette defining file formats.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PairwiseDesignWithAnnotation-new}{\code{PairwiseDesignWithAnnotation$new()}}
\item \href{#method-PairwiseDesignWithAnnotation-generate_translate_dict}{\code{PairwiseDesignWithAnnotation$generate_translate_dict()}}
\item \href{#method-PairwiseDesignWithAnnotation-get_b_labels}{\code{PairwiseDesignWithAnnotation$get_b_labels()}}
\item \href{#method-PairwiseDesignWithAnnotation-get_g_labels}{\code{PairwiseDesignWithAnnotation$get_g_labels()}}
\item \href{#method-PairwiseDesignWithAnnotation-build_file_paths}{\code{PairwiseDesignWithAnnotation$build_file_paths()}}
\item \href{#method-PairwiseDesignWithAnnotation-extract_sample_names}{\code{PairwiseDesignWithAnnotation$extract_sample_names()}}
\item \href{#method-PairwiseDesignWithAnnotation-list_groups_per_batches}{\code{PairwiseDesignWithAnnotation$list_groups_per_batches()}}
\item \href{#method-PairwiseDesignWithAnnotation-list_batches}{\code{PairwiseDesignWithAnnotation$list_batches()}}
\item \href{#method-PairwiseDesignWithAnnotation-find_control_group_per_batches}{\code{PairwiseDesignWithAnnotation$find_control_group_per_batches()}}
\item \href{#method-PairwiseDesignWithAnnotation-get_pairwise_design}{\code{PairwiseDesignWithAnnotation$get_pairwise_design()}}
\item \href{#method-PairwiseDesignWithAnnotation-list_paired_id_per_sample}{\code{PairwiseDesignWithAnnotation$list_paired_id_per_sample()}}
\item \href{#method-PairwiseDesignWithAnnotation-filter_and_set_selected_samples}{\code{PairwiseDesignWithAnnotation$filter_and_set_selected_samples()}}
\item \href{#method-PairwiseDesignWithAnnotation-set_selected_samples}{\code{PairwiseDesignWithAnnotation$set_selected_samples()}}
\item \href{#method-PairwiseDesignWithAnnotation-is_paired}{\code{PairwiseDesignWithAnnotation$is_paired()}}
\item \href{#method-PairwiseDesignWithAnnotation-reset}{\code{PairwiseDesignWithAnnotation$reset()}}
\item \href{#method-PairwiseDesignWithAnnotation-clone}{\code{PairwiseDesignWithAnnotation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesignWithAnnotation-new"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesignWithAnnotation-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new \code{PairwiseDesignWithAnnotation} object.

This method initializes a new \code{PairwiseDesignWithAnnotation} object.
The object represents metadata for expression data, including pairwise
design of experiments and transcriptome annotation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesignWithAnnotation$new(
  pairwise_design_file,
  annotation_file,
  idmapping_file,
  src_dir = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pairwise_design_file}}{Path to the pairwise design file, which must
be in .yml or .csv format.}

\item{\code{annotation_file}}{Path to the annotation file (mapping txid to gid).}

\item{\code{idmapping_file}}{Path to the idmapping file, which allows the
mapping of uniprot id to gene id.}

\item{\code{src_dir}}{Path(s) specified only in case of csv file with the path
for each run to the folder with expression files. In case of several runs
defined in the design file, a named vector is required. Can be specified
for YML also but only 1 value for all runs.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{PairwiseDesignWithAnnotation} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesignWithAnnotation-generate_translate_dict"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesignWithAnnotation-generate_translate_dict}{}}}
\subsection{Method \code{generate_translate_dict()}}{
generate a translate dictionary (named vector) from and id to another
based on the annotation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesignWithAnnotation$generate_translate_dict(from, to)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{from}}{the id type corresponding to the keys}

\item{\code{to}}{the id type corresponding to the values}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Below is precised what is available and id type for value depending
on id type for keys
\itemize{
\item Entrez Gene ID \code{gid} (Or equivalent on Ensembl,UCSC...)
\itemize{
\item Gene symbol \code{symbol}
\item Protein name \code{protein_names}
\item Protein ID \code{uniprot}
\item Taxon ID \code{tax_id}
\item Taxon name \code{tax_name}
\item Typed gene ID \code{tgid}
}
\item RefSeq transcript ID \code{txid} (Or equivalent on Ensembl,UCSC...)
\itemize{
\item Entrez Gene ID \code{gid}
\item rna type \code{type}
\item Taxon ID \code{tax_id}
\item Taxon name \code{tax_name}
\item Typed gene ID \code{tgid}
}
\item Typed gene ID \code{tgid} i.e. gene id with type concatenated (sep is _)
(one gene can have several type, some coding, some not coding)
\itemize{
\item Gene symbol \code{symbol}
\item Protein name \code{protein_names} (warning even if not coding type)
\item Protein ID \code{uniprot} (warning even if not coding type)
\item rna type \code{type}
\item Taxon ID \code{tax_id}
\item Taxon name \code{tax_name}
}
\item Uniprot protein ID \code{uniprot}
\itemize{
\item Entrez Gene ID \code{gid}
\item Protein name \code{protein_names}
\item Gene symbol \code{symbol}
}
}
}

\subsection{Returns}{
a named vector that can translate from an id type to another
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesignWithAnnotation-get_b_labels"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesignWithAnnotation-get_b_labels}{}}}
\subsection{Method \code{get_b_labels()}}{
get batch labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesignWithAnnotation$get_b_labels()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
named vector with batch label (batch code as names)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesignWithAnnotation-get_g_labels"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesignWithAnnotation-get_g_labels}{}}}
\subsection{Method \code{get_g_labels()}}{
get group labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesignWithAnnotation$get_g_labels()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
named vector with group label (group code as names)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesignWithAnnotation-build_file_paths"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesignWithAnnotation-build_file_paths}{}}}
\subsection{Method \code{build_file_paths()}}{
get the paths of files with raw expression data for all samples
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesignWithAnnotation$build_file_paths()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
named vector with paths and sample name as keys
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesignWithAnnotation-extract_sample_names"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesignWithAnnotation-extract_sample_names}{}}}
\subsection{Method \code{extract_sample_names()}}{
get sample names with simple design-based filtering
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesignWithAnnotation$extract_sample_names(
  in_batch = NULL,
  in_group = NULL,
  in_run = NULL,
  basename = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{in_batch}}{A character vector of batch codes. Select or report only
samples in those batches.}

\item{\code{in_group}}{A character vector of group codes. Select or report only
samples in those groups.}

\item{\code{in_run}}{A character vector of run IDs. Select or report only samples
in those runs.}

\item{\code{basename}}{whether to return the basename in place of the complete
name (with run id)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sample names in character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesignWithAnnotation-list_groups_per_batches"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesignWithAnnotation-list_groups_per_batches}{}}}
\subsection{Method \code{list_groups_per_batches()}}{
Get a the groups foreach batch
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesignWithAnnotation$list_groups_per_batches(include_ctrl = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_ctrl}}{whether to include control group}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list of vectors with batch as key and groups as value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesignWithAnnotation-list_batches"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesignWithAnnotation-list_batches}{}}}
\subsection{Method \code{list_batches()}}{
Get a all batchs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesignWithAnnotation$list_batches()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
vector with batch code
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesignWithAnnotation-find_control_group_per_batches"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesignWithAnnotation-find_control_group_per_batches}{}}}
\subsection{Method \code{find_control_group_per_batches()}}{
get the control group per batch
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesignWithAnnotation$find_control_group_per_batches()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
get named vector with batch as key and ctrl group as value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesignWithAnnotation-get_pairwise_design"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesignWithAnnotation-get_pairwise_design}{}}}
\subsection{Method \code{get_pairwise_design()}}{
get the pairwise_design table with design-based filtering
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesignWithAnnotation$get_pairwise_design(
  in_batch = NULL,
  in_group = NULL,
  in_run = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{in_batch}}{A character vector of batch codes. Select or report only
samples in those batches.}

\item{\code{in_group}}{A character vector of group codes. Select or report only
samples in those groups.}

\item{\code{in_run}}{A character vector of run IDs. Select or report only samples
in those runs.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
pairwise_design data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesignWithAnnotation-list_paired_id_per_sample"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesignWithAnnotation-list_paired_id_per_sample}{}}}
\subsection{Method \code{list_paired_id_per_sample()}}{
get the id to be used for paired analysis per sample
only return for samples in batch with a paired design
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesignWithAnnotation$list_paired_id_per_sample()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
named character vector sample_name -> paired_id
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesignWithAnnotation-filter_and_set_selected_samples"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesignWithAnnotation-filter_and_set_selected_samples}{}}}
\subsection{Method \code{filter_and_set_selected_samples()}}{
Select samples based on design. The results will be an intersect with
the previous selection. You can reset the object if it's not desired.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesignWithAnnotation$filter_and_set_selected_samples(
  in_batch = NULL,
  in_group = NULL,
  in_run = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{in_batch}}{A character vector of batch codes. Select or report only
samples in those batches.}

\item{\code{in_group}}{A character vector of group codes. Select or report only
samples in those groups.}

\item{\code{in_run}}{A character vector of run IDs. Select or report only samples
in those runs.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector of the selected samples
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesignWithAnnotation-set_selected_samples"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesignWithAnnotation-set_selected_samples}{}}}
\subsection{Method \code{set_selected_samples()}}{
Set selected samples. The results will be an intersect with
the previous selection. You can reset the object if it's not desired.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesignWithAnnotation$set_selected_samples(sample_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample_list}}{vector of sample codes to select}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector of the selected samples
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesignWithAnnotation-is_paired"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesignWithAnnotation-is_paired}{}}}
\subsection{Method \code{is_paired()}}{
get the list of batch with a paired design (DEPRECATED)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesignWithAnnotation$is_paired()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesignWithAnnotation-reset"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesignWithAnnotation-reset}{}}}
\subsection{Method \code{reset()}}{
Reset object, this affects:
\itemize{
\item samples selection (as initially)
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesignWithAnnotation$reset()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
vector of the selected samples
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseDesignWithAnnotation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseDesignWithAnnotation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseDesignWithAnnotation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
