% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_PairwiseLFC.r
\name{PairwiseLFC}
\alias{PairwiseLFC}
\title{Class representing raw Log Fold change results from a pairwaise design}
\description{
This class generates raw Log Fold change without statistical inference or
tests
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PairwiseLFC-new}{\code{PairwiseLFC$new()}}
\item \href{#method-PairwiseLFC-get_results}{\code{PairwiseLFC$get_results()}}
\item \href{#method-PairwiseLFC-clone}{\code{PairwiseLFC$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseLFC-new"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseLFC-new}{}}}
\subsection{Method \code{new()}}{
Initialize \code{PairwiseLFC} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseLFC$new(
  expr_data,
  ncpus = 1,
  method_ctrl = "ctrl_med",
  compute_chunk = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expr_data}}{ExprData : expression data}

\item{\code{ncpus}}{number of cpus to use for computation (default = 1)}

\item{\code{method_ctrl}}{method to compute the reference expression
\itemize{
\item \code{ctrl_each} : samples are compared to each ctrl
\item \code{ctrl_med} : samples are compared to the median of ctrl
\item \code{ctrl_mean} : samples are compared to the mean of ctrl
\item \code{ctrl_geo} : samples are compared to the geo mean of ctrl
\item \code{ctrl_geo_mod} : samples are compared to the modified geo mean of ctrl
}}

\item{\code{compute_chunk}}{a vector of index to select the groups to compute
this option is present to limit compution time.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{PairwiseLFC} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseLFC-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseLFC-get_results}{}}}
\subsection{Method \code{get_results()}}{
Get LFC results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseLFC$get_results()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A nested dataframe.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseLFC-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseLFC-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseLFC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
