% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_plot_complex.r
\name{plot_complex_check}
\alias{plot_complex_check}
\title{Helper function to check plot complex arguments}
\usage{
plot_complex_check(plot_scale, inter_norm, inter_norm_fact_opts, in_batch)
}
\arguments{
\item{plot_scale}{Scale for plotting ("design", "batch", or "group")}

\item{inter_norm}{Whether inter normalization is applied}

\item{inter_norm_fact_opts}{Inter normalization options}

\item{in_batch}{Batch filter}
}
\value{
NULL (stops execution if checks fail)
}
\description{
Helper function to check plot complex arguments
}
\keyword{internal}
