% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_plot.r
\name{plot_dist_per_sample_helper}
\alias{plot_dist_per_sample_helper}
\title{Helper function for ExprData$plot_dist_per_sample}
\usage{
plot_dist_per_sample_helper(
  data,
  design,
  geoms = c("boxplot"),
  mean_fun = NULL,
  fn_str
)
}
\arguments{
\item{data}{data frame with columns: batch, group, sample, value}

\item{design}{private$design see ExprData}

\item{geoms}{see ExprData$plot_dist_per_sample}

\item{mean_fun}{see ExprData$plot_dist_per_sample}

\item{fn_str}{string representation of y-axis transform}
}
\value{
ggplot2 graph
}
\description{
Helper function for ExprData$plot_dist_per_sample
}
\keyword{internal}
