% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_plot.r
\name{plot_sum_per_type_helper}
\alias{plot_sum_per_type_helper}
\title{Helper function for ExprData$plot_sum_per_type_per_sample}
\usage{
plot_sum_per_type_helper(
  data,
  design,
  same_scale = TRUE,
  horizontal_bar = TRUE,
  log2_expr = FALSE,
  exclude_type = c()
)
}
\arguments{
\item{data}{data frame with columns: batch, group, sample, type, total}

\item{design}{private$design see ExprData}

\item{same_scale}{see ExprData$plot_sum_per_type_per_sample}

\item{horizontal_bar}{see ExprData$plot_sum_per_type_per_sample}

\item{log2_expr}{see ExprData$plot_sum_per_type_per_sample}

\item{exclude_type}{see ExprData$plot_sum_per_type_per_sample}
}
\value{
ggplot2 graph
}
\description{
Helper function for ExprData$plot_sum_per_type_per_sample
}
