% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_summary.r
\name{plot_summarized_etags}
\alias{plot_summarized_etags}
\title{Helper function for ExprData$sum_per_type_per_sample}
\usage{
plot_summarized_etags(results, design)
}
\arguments{
\item{results}{see summarize_etags}

\item{design}{private$design}
}
\value{
list of plots
}
\description{
Helper function for ExprData$sum_per_type_per_sample
}
