% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_plot_complex.r
\name{prepare_tag_selection}
\alias{prepare_tag_selection}
\title{Helper function to prepare tag selection}
\usage{
prepare_tag_selection(
  main_etag,
  annotation,
  selected_ids,
  tags = NULL,
  tag_type = NULL
)
}
\arguments{
\item{main_etag}{Main expression tag ID}

\item{annotation}{Annotation object}

\item{selected_ids}{Currently selected IDs}

\item{tags}{Vector of tag ids to use for the plot}

\item{tag_type}{Type of tag to use}
}
\value{
Vector of selected tag IDs
}
\description{
Helper function to prepare tag selection
}
\keyword{internal}
