% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{safe_translate_ids}
\alias{safe_translate_ids}
\title{Safely translate IDs using a named vector}
\usage{
safe_translate_ids(named_vector, vector_to_translate)
}
\arguments{
\item{named_vector}{A named vector where names are the original IDs and
values are the translated IDs.}

\item{vector_to_translate}{A character vector of IDs to translate.}
}
\value{
A character vector with translated IDs, preserving original IDs
where translation was not possible.
}
\description{
Translates a vector of IDs using a named vector (dictionary). If an ID
is not found in the named vector (resulting in NA), the original ID is
kept.
}
\examples{
translation_map <- c("ID1" = "NewID1", "ID2" = "NewID2")
ids_to_translate <- c("ID1", "ID3", "ID2")
safe_translate_ids(translation_map, ids_to_translate)
# Should return: c("NewID1", "ID3", "NewID2")
}
