% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_norm.r
\name{set_mean_function}
\alias{set_mean_function}
\title{Helper function to set mean function based on method name}
\usage{
set_mean_function(method_name)
}
\arguments{
\item{method_name}{Method name for mean calculation}
}
\value{
Function for computing mean
}
\description{
Helper function to set mean function based on method name
}
