% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{simple_chord}
\alias{simple_chord}
\title{Generate a simple chord plot}
\usage{
simple_chord(
  data,
  superset_colname = "batch",
  set_colname = "group",
  data_colname = "data",
  data_list_colname = "gid",
  scale_groups = TRUE,
  isolate_supersets = TRUE
)
}
\arguments{
\item{data}{a tibble data frame with nested gene list}

\item{superset_colname}{name of the column with the superset code}

\item{set_colname}{name of the column with the set code}

\item{data_colname}{name of the column with the nested lists}

\item{data_list_colname}{name of the subcolumn within the nested lists that
hold the ids of interest.}

\item{scale_groups}{whether to scale the sizes of groups on the graph}

\item{isolate_supersets}{whether to isolate the sets within the same
superset}
}
\description{
Generate a simple chord plot
}
