% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_summary.r
\name{sum_per_type_per_sample_helper}
\alias{sum_per_type_per_sample_helper}
\title{Helper function for ExprData$sum_per_type_per_sample}
\usage{
sum_per_type_per_sample_helper(
  annotation,
  main_etag,
  design,
  get_raw_data_callback,
  compute_norm_callback,
  intra_norm = FALSE,
  log2_expr = FALSE
)
}
\arguments{
\item{annotation}{private$annotation (see ExprData)}

\item{main_etag}{private$main_etag (see ExprData)}

\item{design}{private$design (see ExprData)}

\item{get_raw_data_callback}{A function to get the raw data}

\item{compute_norm_callback}{A function to compute the normalized data}

\item{intra_norm}{see ExprData$sum_per_type_per_sample}

\item{log2_expr}{see ExprData$sum_per_type_per_sample}
}
\value{
data frame with counts with columns type, sample, total
}
\description{
Helper function for ExprData$sum_per_type_per_sample
}
