% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_summary.r
\name{summarize_etags}
\alias{summarize_etags}
\title{Helper function for ExprData$show_etags_summary}
\usage{
summarize_etags(selected_ids, annotation, at_gene_level, type = "etags", data)
}
\arguments{
\item{selected_ids}{private$selected_ids (see ExprData)}

\item{annotation}{private$annotation (see ExprData)}

\item{at_gene_level}{private$at_gene_level (see ExprData)}

\item{type}{see ExprData$show_etags_summary}

\item{data}{input data}
}
\value{
A list of count tables per feature
}
\description{
Helper function for ExprData$show_etags_summary
}
