% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\docType{data}
\name{THEME_NEXOMIS}
\alias{THEME_NEXOMIS}
\title{ggplot2 theme for nexomis}
\format{
An object of class \code{theme} (inherits from \code{ggplot2::theme}, \code{gg}, \code{S7_object}) of length 144.
}
\usage{
THEME_NEXOMIS
}
\description{
Simple theme
}
\keyword{datasets}
