% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\docType{data}
\name{UP_COLOR}
\alias{UP_COLOR}
\title{Color palette for nexomis: Up red}
\format{
An object of class \code{character} of length 1.
}
\usage{
UP_COLOR
}
\description{
Color palette for nexomis: Up red
}
\keyword{datasets}
