% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\docType{data}
\name{UP_TO_DOWN_GRADIENT_COLORS}
\alias{UP_TO_DOWN_GRADIENT_COLORS}
\title{Color palette gradient for nexomis
from UP red to DOWN blue through MID yellow-white}
\format{
An object of class \code{character} of length 11.
}
\usage{
UP_TO_DOWN_GRADIENT_COLORS
}
\description{
Color palette gradient for nexomis
from UP red to DOWN blue through MID yellow-white
}
\keyword{datasets}
