% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_norm.r
\name{compute_norm_fact_helper}
\alias{compute_norm_fact_helper}
\title{Helper function for ExprData$compute_norm_fact}
\usage{
compute_norm_fact_helper(
  selected_ids,
  design,
  intra_norm_fact,
  inter_norm_fact,
  inter_norm_fact_opts,
  in_batch = NULL,
  in_group = NULL,
  inter_norm = FALSE,
  intra_norm = TRUE
)
}
\arguments{
\item{selected_ids}{private$selected_ids (see ExprData)}

\item{design}{private$design (see ExprData)}

\item{intra_norm_fact}{private$intra_norm_fact (see ExprData)}

\item{inter_norm_fact}{private$inter_norm_fact (see ExprData)}

\item{inter_norm_fact_opts}{private$inter_norm_fact_opts (see ExprData)}

\item{in_batch}{see ExprData$compute_norm_fact}

\item{in_group}{see ExprData$compute_norm_fact}

\item{inter_norm}{see ExprData$compute_norm_fact}

\item{intra_norm}{see ExprData$compute_norm_fact}
}
\value{
normalization factor matrix
}
\description{
Helper function for ExprData$compute_norm_fact
}
