% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_plot_complex.r
\name{create_prcomp_plot}
\alias{create_prcomp_plot}
\title{Helper function to create PCA plots}
\usage{
create_prcomp_plot(
  in_data,
  in_title,
  data_design,
  pca_plot_dims = c(1, 2),
  mshape = "batch",
  mcolor = "group",
  prcomp_args = list(),
  ggplot_mod = NULL,
  color_palette = "BrBG",
  point_size = 8
)
}
\arguments{
\item{in_data}{Expression data matrix}

\item{in_title}{Plot title}

\item{data_design}{Design data frame}

\item{pca_plot_dims}{Integer vector specifying which principal components to plot}

\item{mshape}{Character string specifying variable for point shapes}

\item{mcolor}{Character string specifying variable for point colors}

\item{prcomp_args}{Arguments for prcomp}

\item{ggplot_mod}{Additional ggplot modifications}

\item{color_palette}{Color palette name}

\item{point_size}{Numeric value for point size (default: 8)}
}
\value{
ggplot object
}
\description{
Helper function to create PCA plots
}
\keyword{internal}
