% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_id_mapping.r
\name{fetch_id_mapping}
\alias{fetch_id_mapping}
\title{Extract id_mapping table from UniProt API}
\usage{
fetch_id_mapping(tax_id)
}
\arguments{
\item{tax_id}{Taxonomic ID to filter the UniProt database.}
}
\value{
A data frame containing the id_mapping table.
}
\description{
This function extracts the id_mapping table from the UniProt API based on the specified taxonomic ID.
}
\examples{
\dontrun{
id_mapping <- fetch_id_mapping(tax_id = 9606)
}

}
