% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExprData_utils_summary.r
\name{summarize_etags}
\alias{summarize_etags}
\title{Helper function for ExprData$show_etags_summary}
\usage{
summarize_etags(annotation, at_gene_level, data, tr_fn, sum_fn)
}
\arguments{
\item{annotation}{private$annotation (see ExprData)}

\item{at_gene_level}{private$at_gene_level (see ExprData)}

\item{data}{input data}

\item{tr_fn}{transformation function to apply to data}

\item{sum_fn}{aggregation function to apply after transformation}
}
\value{
A list of count tables per feature
}
\description{
Helper function for ExprData$show_etags_summary
}
\keyword{internal}
