% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\docType{data}
\name{THEME_NEXOMIS}
\alias{THEME_NEXOMIS}
\title{ggplot2 theme for nexomis}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 136.
}
\usage{
THEME_NEXOMIS
}
\description{
Simple theme
}
\keyword{datasets}
